Update AT3_answers.Rmd
Question 6: We need to run the command "colnames" to show the column names as required. 
Question 7: In my opinion, we don't need to calculate separately on each site. The question said "at both sites".
Question 8: The question is "Make a box plot of tree circumference at the start and end of the study at both sites.". It means we must make a box plot of tree circumference in 2004 and 2019, not 2004 and 2014.